requirements/base.txt
aio-pika==6.8.0
aiogram==2.12.1
aiohttp==3.7.4.post0
aioredis==1.3.1
aiormq==3.3.1
amqp==5.0.5
apns2==0.7.2
APScheduler==3.6.3
aredis==1.1.8
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.0
billiard==3.6.3.0
blessings==1.7
bpython==0.21
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.6
curtsies==0.3.5
cwcwidth==0.1.4
daphne==3.0.1
Django==3.1.7
django-cors-headers==3.7.0
django-push-notifications==2.0.0
django-solo==1.1.5
djangorestframework==3.12.2
five==0.4.0
greenlet==1.0.0
h2==2.6.2
hiredis==1.1.0
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kombu==5.0.2
lxml==4.6.3
msgpack==1.0.2
multidict==5.1.0
numpy==1.20.2
pamqp==2.3.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.6
py-vapid==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-telegram-bot==13.4.1
pytz==2021.1
pywebpush==1.13.0
pyxdg==0.27
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zope.interface==5.3.0
